+++ src/containerapp/azext_containerapp/custom.py
@@ -179,7 +179,7 @@ def update_containerapp_yaml(cmd, name, resource_group_name, file_name, from_rev
 
     if not containerapp_def:
         raise ValidationError("The containerapp '{}' does not exist".format(name))
-
+    existed_environment_id = containerapp_def['properties']['environmentId']
     containerapp_def = None
 
     # Deserialize the yaml into a ContainerApp object. Need this since we're not using SDK
@@ -224,6 +224,10 @@ def update_containerapp_yaml(cmd, name, resource_group_name, file_name, from_rev
         if "template" not in containerapp_def["properties"]:
             containerapp_def["properties"]["template"] = {}
         containerapp_def["properties"]["template"]["revisionSuffix"] = None
+
+    # Remove the environmentId in the PATCH payload if the environment not been changed
+    if containerapp_def["properties"].get('environmentId') and containerapp_def["properties"].get('environmentId') == existed_environment_id:
+        del containerapp_def["properties"]['environmentId']
 
     try:
         r = ContainerAppClient.update(